+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 <!--next-version-placeholder-->
 
+## v0.1.2 (23/09/2023)
+
+- In this version, cplex package is used instead of gurobipy for solving
+maxcut problem to improve the runtime of maxcut solver.
+
 ## v0.1.1 (23/09/2023)
 
 - Fixes an important bug when adding more data to fit the model again.